@@ -3,7 +3,7 @@ package uk.gov.hmcts.reform.bulkscan.payment.processor.client.processor.response
 public class PaymentStatusReponse {
     private String status;
 
-    private PaymentStatusReponse(){
+    private PaymentStatusReponse() {
     }
 
     public PaymentStatusReponse(String status) {
